% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_estimator.R
\name{tidy_estimator}
\alias{tidy_estimator}
\title{Tidy Results from a List of Estimators}
\usage{
tidy_estimator(..., vcov = NULL, alpha = 0.05, null = 0, term = NULL)
}
\arguments{
\item{...}{A named list of fitted model objects. Each model is passed to `marginaleffects::hypotheses()`.}

\item{vcov}{An optional variance-covariance matrix or a function to compute it. Passed to `tidy_estimator`.}

\item{alpha}{A numeric value between 0 and 1 specifying the significance level for confidence intervals (default is 0.05).}

\item{null}{A numeric value representing the null hypothesis for the parameter estimates (default is 0).}

\item{term}{An optional character vector of terms to filter the output. If specified, only rows with matching terms are returned.}
}
\value{
A data frame combining the tidied results from all models in the list.
  Columns include:
\describe{
  \item{estimator}{The label for the estimator (corresponding to the list names).}
  \item{term}{The name of the parameter (e.g., model term).}
  \item{estimate}{The estimated value of the parameter.}
  \item{p.value}{The p-value for the hypothesis test of the parameter.}
  \item{conf.low}{The lower bound of the confidence interval.}
  \item{conf.high}{The upper bound of the confidence interval.}
}
}
\description{
This function applies extracts a "tidy" output from a fitted model, or from a named list of fitted models.
The information is extracted using the `hyotheses()` function from the `marginaleffects` package.
}
\examples{
library(SimpleDesign)
model1 <- lm(mpg ~ wt, data = mtcars)
model2 <- lm(mpg ~ disp, data = mtcars)
models <- list(OLS_wt = model1, OLS_disp = model2)
tidy_estimator_list(models, alpha = 0.05, null = 0)

}
